+++ src/services/Elastic.Documentation.Services/Changelog/ChangelogConfiguration.cs
@@ -2,7 +2,7 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-namespace Elastic.Documentation.Services.ReleaseNotes;
+namespace Elastic.Documentation.Services.Changelog;
 
 /// <summary>
 /// Configuration for changelog generation
